// client/src/containers/interface.ts
type RowType = {
  key: number;
  first_name: string;
  last_name: string;
  email: string;
  course_name: string;
  passing_year: string;
  university_name: string;
  department: string;
  job_title: string;
  company_name: string;
  contact_no: string;
  edit: string;
};

export type { RowType };